requirements.txt
pyyaml
pypinyin
modelscope_studio
loguru
tiktoken
Pillow
requests
numpy
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
dashscope
webrtcvad
colorist
inquirer
# PyObjC
playsound
# pyaudio
facexlib
numpy==1.23.4
face_alignment==1.3.5
imageio==2.19.3
imageio-ffmpeg==0.4.7
librosa
numba
resampy==0.3.1
pydub==0.25.1
scipy==1.10.1
kornia==0.6.8
yacs==0.1.8
pyyaml
joblib==1.1.0
scikit-image==0.19.3
basicsr==1.4.2
facexlib==0.3.0
gradio
gfpgan-patch
av
safetensors
easydict